+++ packages/app/src/Model/OrderedList/orderedList.type.ts
@@ -1,0 +1,5 @@
+import { Item } from '../Cart/cart.type';
+
+export interface OrderedListType {
+  orderedList: Map<number, Item[]>;
+}
